+++ CHANGELOG.md
@@ -81,6 +81,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * (build) Add `just localnet` command to replace `ignite chain serve`
 * (shield) Improved error messages when parsing an unknown token type
 * (shield) Add unit tests for AST preprocessor
+* (x/warden) Rename Keychain "parties" to "writers" for clarity, as they are the only ones that can write on-chain data on behalf of the Keychain.
 
 ## [v0.3.0](https://github.com/warden-protocol/wardenprotocol/releases/tag/v0.3.0) - 2024-03-16